pyproject.toml
[project]
name = "perception-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "decord>=0.6.0",
    "einops>=0.8.1",
    "ftfy>=6.3.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
